+++ .github/workflows/orion/add_uncompleted_experiment.py
@@ -1,6 +1,7 @@
 """
 Helper script to add an uncompleted experiment to
 db_dashboard_full_with_uncompleted_experiment.pkl
+built from a copy of db_dashboard_full.pkl
 """
 from datetime import datetime, timedelta
